% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flooded-package.R
\docType{package}
\name{flooded-package}
\alias{flooded}
\alias{flooded-package}
\title{flooded: Portable Floodplain Delineation from DEM and Stream Network}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Delineate floodplains from digital elevation models and stream networks using the Valley Confinement Algorithm (VCA). Combines slope thresholding, cost-distance analysis, and flood surface modelling to identify unconfined valleys. Ported from the bcfishpass lateral habitat pipeline (Python) with origins in the USDA VCA Toolbox and BlueGeo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewGraphEnvironment/flooded}
  \item \url{https://newgraphenvironment.github.io/flooded/}
  \item Report bugs at \url{https://github.com/NewGraphEnvironment/flooded/issues}
}

}
\author{
\strong{Maintainer}: Allan Irvine \email{al@newgraphenvironment.com} (\href{https://orcid.org/0000-0002-3495-2128}{ORCID})

Other contributors:
\itemize{
  \item Devin Cairns (Valley Confinement Algorithm (MIT license, Blue Geosimulation)) [copyright holder]
  \item Simon Norris (bcfishpass lateral habitat assembly (Apache 2.0)) [copyright holder]
  \item Canadian Wildlife Federation (bcfishpass (Apache 2.0)) [copyright holder]
  \item Province of British Columbia (bcfishpass (Apache 2.0)) [copyright holder]
  \item New Graph Environment (bcfishpass (Apache 2.0)) [copyright holder]
}

}
\keyword{internal}
