% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_flood_trim.R
\name{fl_flood_trim}
\alias{fl_flood_trim}
\title{Subtract mask layers from a binary raster}
\usage{
fl_flood_trim(x, ...)
}
\arguments{
\item{x}{A \code{SpatRaster} with binary values (\code{0}/\code{1}).}

\item{...}{One or more \code{SpatRaster} exclusion masks. Cells with value \code{1}
in any mask are removed from \code{x}.}
}
\value{
A \code{SpatRaster} with the same grid as \code{x}. Cells overlapping any
exclusion mask are set to \code{0}.
}
\description{
Removes (sets to \code{0}) cells in a binary raster that overlap with any
exclusion mask. Useful for trimming floodplain by urban areas, steep
terrain, waterbodies, or other features.
}
